+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - form id setting
 - bug in editing splitted data
 - database creation
+- error when ' in element name
+- reordering of elements
 
 ### Updated